add: maps marker

+++ app/src/main/java/com/bangkit/storyapp/data/model/response/GetStoriesResponse.kt
@@ -17,6 +17,6 @@ data class ListStory(
     val description: String,
     val photoUrl: String,
     val createdAt: String,
-    val lat: Float?,
-    val lon: Float?
+    val lat: Double?,
+    val lon: Double?
 )
